% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.r
\name{dnamalci.get.model}
\alias{dnamalci.get.model}
\title{Retrieve model details}
\usage{
dnamalci.get.model(name)
}
\arguments{
\item{name}{Name of the model.}
}
\value{
Model details formatted as a named list.
}
\description{
Retrieve model details
}
