% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.r
\name{dnamalci.load.test.data}
\alias{dnamalci.load.test.data}
\title{Retrieve dna methylation test dataset}
\usage{
dnamalci.load.test.data()
}
\description{
This dataset is a minimal subset of CpGs from the 
Tsaprouni et al. 2014 data available at the
Gene Expression Omnibus (GEO), accession number GSE50660:
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE50660
}
