% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.r
\name{dnamalci.score}
\alias{dnamalci.score}
\title{Calculate any of the DNA methylation alcohol scores from Liu et al. 2016}
\usage{
dnamalci.score(x, model = "dnamalc.144cpg")
}
\arguments{
\item{x}{DNA methylation matrix (rows=CpG sites, columns=samples).}

\item{model}{Model name from available list (\code{\link{dnamalci.models}()}).
Default is the `dnamalc.144cpg` score.}
}
\value{
Scores derived for each sample in \code{x}.
}
\description{
Calculate any of the DNA methylation alcohol scores from Liu et al. 2016
}
